==> 01.Hooklog_Prerocess.ipynb <==
build standard ans of malware family

import pandas as pd
import numpy as np
from collections import Counter
from tqdm import tqdm
import os, shutil

ori_df = pd.read_csv("./data/Hooklog/VTRepo_wn_AAMerge.csv")
ori_df = ori_df.fillna("no_name")
ori_df

ori_df[ori_df['Unnamed: 0'] == 'fed55bd9dd509d59d240700404c38d20de75e1e72218c83e473e6833cb6eb5b3']

ans_df = pd.DataFrame(columns=["AV_hash","fam_name"]) #原始query的hash
for i in tqdm(range(len(ori_df))):
    ori_li = ori_df.iloc[i].values.tolist()
    hash_id = ori_li[0]
    del ori_li[0]
    name_li = [x for x in ori_li if x != "no_name"]
    name_count = Counter(name_li).most_common()
    try:
        first_num = name_count[0][1]
    except IndexError :
        continue
    try:
        second_num = name_count[1][1]
    except IndexError:
        second_num = int((len(ori_df.columns) - 1)/6) #all AV vendors number 
    if first_num > 2*second_num: #1/3 of AV vender agree if there's no second_num
        temp = [hash_id,name_count[0][0]]
        temp = pd.Series(temp, index = ans_df.columns)
        ans_df = ans_df.append(temp, ignore_index=True)
ans_df

ans_df[ans_df['AV_hash'] == 'fca30fe156387a28e5aad03ec4fa945e828f9abd2760b074ce535f803f26185a']

hkl_dir = './data/Hooklog/trace_original_all_param/'
all_hkls_li = next(os.walk(hkl_dir))[2]
print(len(all_hkls_li))

ans2_df = pd.DataFrame(columns=["pid","fam_name"]) #成功有錄製的hash
ans3_df = pd.DataFrame(columns=["success_hash","fam_name"])
processed_hash = []
needed_hash_li = set(ans_df.AV_hash.tolist())
for hkl in tqdm(all_hkls_li):
    hk_hash = hkl.split("_")[0]
    if hk_hash in needed_hash_li:
        temp = [hkl,ans_df[ans_df.AV_hash == hk_hash]['fam_name'].values[0]]
        temp = pd.Series(temp, index = ans2_df.columns)
        ans2_df = ans2_df.append(temp , ignore_index=True)
        if hk_hash not in processed_hash:
            temp = [hk_hash,ans_df[ans_df.AV_hash == hk_hash]['fam_name'].values[0]]
            temp = pd.Series(temp, index= ans3_df.columns)
            ans3_df = ans3_df.append(temp, ignore_index=True)
            processed_hash.append(hk_hash)
ans3_df     #10728 processes

def basic_statistics(all_length):
    '''
    input: length list of elements e.g.[1,1,1,3,5,9,4,2,1,3,54,78,5...]
    output1: mean、std、mode、min、q1、median(q2)、q3、max、iqr、outlier、far out
    output2: statistics graph、10%~90% form
    '''
    stat_dict = {}
    stat_dict['mean'] = np.mean(all_length)
    stat_dict['std'] = np.std(all_length)
    stat_dict['mode'] = np.argmax(np.bincount(all_length))
    stat_dict['min'] = np.min(all_length)
    stat_dict['q1'] = np.quantile(all_length,0.25)
    stat_dict['median'] = np.quantile(all_length,0.5)
    stat_dict['q3'] = np.quantile(all_length,0.75)
    stat_dict['max'] = np.max(all_length)
    stat_dict['iqr'] = stat_dict['q3'] - stat_dict['q1']
    stat_dict['outlier'] = stat_dict['q3'] + 1.5*stat_dict['iqr']
    stat_dict['far_out'] = stat_dict['q3'] + 3*stat_dict['iqr']
    for i in [10,20,30,40,50,60,70,80,90,100]:
        stat_dict[str(i)+'%'] = np.percentile(all_length,i)
    return pd.DataFrame.from_dict(stat_dict,orient='index',columns=['length'])

statistics = ans_df.fam_name.tolist()
statistics_pd = pd.DataFrame.from_dict(dict(Counter(statistics)),columns=['samples_num'],orient='index')
print("total fam_number:",len(statistics_pd))
basic_statistics(statistics_pd.samples_num.tolist())

statistics2 = ans2_df.fam_name.tolist()
statistics2_pd = pd.DataFrame.from_dict(dict(Counter(statistics2)),columns=['processes_num'],orient='index')
print("total fam_number:",len(statistics2_pd))
basic_statistics(statistics2_pd.processes_num.tolist())

statistics3 = ans3_df.fam_name.tolist()
statistics3_pd = pd.DataFrame.from_dict(dict(Counter(statistics3)),columns=['profiled_samples_num'],orient='index')
print("total fam_number:",len(statistics3_pd))
basic_statistics(statistics3_pd.profiled_samples_num.tolist())

statistics_merge_pd = pd.merge(statistics_pd,statistics3_pd,right_index=True,left_index=True,how='outer')
statistics_merge_pd = pd.merge(statistics_merge_pd, statistics2_pd, right_index=True, left_index=True,how='outer')
statistics_merge_pd = statistics_merge_pd.fillna(0)
statistics_merge_pd = statistics_merge_pd.astype('int')
statistics_merge_pd.to_excel('./results/Hooklog/statistics_fam_dist.xlsx')
statistics_merge_pd

def family_take(se):
    if se infam_names:
        return 'take'
    else:
        return 'not_take'

### pick family

used_df = statistics_merge_pd[(statistics_merge_pd['processes_num']>160) & (statistics_merge_pd['processes_num']<286)]
fam_names = used_df.index.tolist()
print(fam_names, len(fam_names),'families')
def family_take(se):
    if se in fam_names:
        return 'take'
    else: 
        return 'not_take'
ans2_df['take'] = ans2_df.fam_name.map(family_take)
final_df = ans2_df[ans2_df['take'] == 'take']
final_df.reset_index(inplace=True,drop=True)
final_df.drop(['take'],inplace=True,axis=1)
final_df

pick_dir = "./data/Hooklog/trace_picked_fam/"
for i in range(len(final_df)):
    file_name = final_df.loc[i,'pid']
    fam_name = final_df.loc[i,'fam_name']
    out_dir = pick_dir+fam_name+'/'
    if not os.path.exists(out_dir):
        os.makedirs(out_dir,exist_ok=True)
    shutil.copyfile(hkl_dir + file_name,out_dir+file_name)
