==> 1.AnalyzeData-Copy1.ipynb <==
import os, sys, glob, csv
# os.environ["CUDA_DEVICE_ORDER"]="PCI_BUS_ID"
# The GPU id to use, usually either "0" or "1"
os.environ["CUDA_VISIBLE_DEVICES"]="1" 
import numpy as np
import pandas as pd
import shutil
import pickle
import tqdm
from tqdm import tqdm
import xlsxwriter
import random
from random import shuffle
from math import log, floor
import re
import collections
from collections import Counter
import string
import unicodedata as udata
# import pause, datetime
import matplotlib.pyplot as plt
import seaborn as sns
from pathlib import Path
from distutils.dir_util import copy_tree
import sklearn
from sklearn.metrics import *
import itertools as it
import nltk
from nltk.corpus import stopwords
from nltk.tokenize import wordpunct_tokenize
from nltk import word_tokenize
from nltk.stem.porter import *
from nltk.tokenize import RegexpTokenizer
from collections import Counter
import functools
import spacy

import tensorflow.keras.preprocessing.text as T

stop words

stop_words_1 = (stopwords.words('english')) #Stopword
with open('data/preprocess/stop_words.txt') as f:
    stop_words_2 = f.read().splitlines() #stop_list1
stop_words_3 = pickle.load(open('data/preprocess/stop_list2.pkl','rb'))
stop_words_all = set(stop_words_1 + stop_words_2 + stop_words_3)
len(stop_words_all) # stopwords#

sp_tokenizer = spacy.load("en")
kk = sp_tokenizer("It's (beautiful) 4toys doesn't")
token_li=[]
for token in kk:
    token_li.append(token.text)
token_li

Prepare preprocess

sp_tokenizer = spacy.load("en")
dil= r"[!\"#$%&()*+,-./:;<=>?@[\\]^_`{|}~]+\ *" 
dil_filter = '!\"#$%&()*+,-./:;<=>?@[\\]^_\`{|}~\t\n'
glove = pd.read_table('./data/preprocess/glove.840B.300d.txt', sep=" ", index_col=0, header=None, quoting=csv.QUOTE_NONE)
glove_words = set(glove.index.tolist())
len(glove_words)

以下deprecated

def text_preprcess(text,tokenizers='nltk'):
    '''
    Input: string
    Return: preprocessed list
    '''
    tokens = [i for i in text.split() if i not in stop_words_all]
    tokens_str = ' '.join(tokens)
    if tokenizers == 'keras':
        tokens = T.text_to_word_sequence(tokens_str,filters=dil_filter) # lower casse & tokenize
        tokens = list(filter(None, tokens))
    elif tokenizers == 'nltk':
#         tokens = re.sub(dil," ",tokens_str.lower())
        tokens = [i for i in tokens_str.lower() if i not in dil_filter]
        tokens = ''.join(tokens)
        tokens = word_tokenize(tokens)
        tokens = list(filter(None, tokens))
#         tokens = [x.lower() for x in tokens]
    elif tokenizers == 'spacy':
#         tokens = re.sub(dil," ",tokens_str.lower())
        tokens = [i for i in tokens_str.lower() if i not in dil_filter]
        tokens = ''.join(tokens)
        tokens = sp_tokenizer(tokens)
        token_li = []
        for token in tokens:
            token_li.append(token.text)
        tokens = list(filter(None, token_li))
#         tokens = [x.lower() for x in token_li]
    tokens_li = []
    for token in tokens:
        tokens_li.append(''.join([i for i in token if not i.isdigit()]))
    tokens_str = ' '.join(tokens_li)
    tokens_str = re.sub(dil," ",tokens_str.lower())
    tokens_li = tokens_str.split()
    tokens_li = list(filter(None, tokens_li))
    ori_tokens_num = len(tokens_li)
    tokens_li = [i for i in tokens_li if i in glove_words]
    try:
        rate = len(tokens_li)/ori_tokens_num
    except ZeroDivisionError:
        rate = np.nan
    return tokens_li , rate#ori_tokens_num/len(tokens_li)
        
    

# train_df = pd.DataFrame(columns=['id','category','text','original_tokenize_num','preprocess_text','preprocss_tokenize_num'])
train_df = pd.DataFrame(columns=['id','category','text','original_tokenize_num',
                                 'text_keras','rate_keras','text_nltk','rate_nltk','text_apacy','rate_spacy'])
train_dir = './data/20news-bydate-v3/20news-bydate-train/'
cat_dir = next(os.walk(train_dir))[1]
count = 0
for cat in tqdm(cat_dir):
    in_dir = train_dir + cat + '/'
    news_list = next(os.walk(in_dir))[2]
    news_list = [os.path.join(in_dir, f) for f in news_list]
    news_list = list(filter(lambda f: f.endswith(".txt"), news_list))
    for news_path in news_list:
        id_num = news_path.split('/')[-1].split('.')[0]
        with open(news_path,'r',encoding='latin1') as f:
            news_text = f.read()
            news_text = re.sub('\n',' ',news_text)
            news_text = re.sub('\t',' ',news_text)
            news_text_tok_ori = T.text_to_word_sequence(news_text,filters='',lower=False)
            news_text_ori = " ".join(news_text_tok_ori)
            news_text_tok_num_ori = len(news_text_tok_ori)
            news_text_tok_keras,rate_keras = text_preprcess(news_text,tokenizers='keras')
            news_text_tok_nltk,rate_nltk = text_preprcess(news_text,tokenizers='nltk')
            news_text_tok_spacy,rate_spacy = text_preprcess(news_text,tokenizers='spacy')
        train_df.loc[count] = [id_num,cat,news_text_ori,news_text_tok_num_ori,
                               news_text_tok_keras,rate_keras,news_text_tok_nltk,rate_nltk,news_text_tok_spacy,rate_spacy]
        count+=1
#         break
    
# train_df.loc[0] = [0,0,0]
print(train_df.rate_keras.mean() , train_df.rate_nltk.mean(),train_df.rate_spacy.mean()  )
train_df

* keras tokenizer: 95.95% tokens in glove
* nltk tokenizer: 95.96% tokens in glove
* spacy tokenizer: 96.05% tokens in glove

train_df = pd.DataFrame(columns=['id','category','text','original_tokenize_num','preprocess_text','preprocss_tokenize_num'])
train_dir = './data/20news-bydate-v3/20news-bydate-train/'
cat_dir = next(os.walk(train_dir))[1]
count = 0
for cat in tqdm(cat_dir):
    in_dir = train_dir + cat + '/'
    news_list = next(os.walk(in_dir))[2]
    news_list = [os.path.join(in_dir, f) for f in news_list]
    news_list = list(filter(lambda f: f.endswith(".txt"), news_list))
    for news_path in news_list:
        id_num = news_path.split('/')[-1].split('.')[0]
        with open(news_path,'r',encoding='latin1') as f:
            news_text = f.read()
            news_text = re.sub('\n',' ',news_text)
            news_text_tok_ori = T.text_to_word_sequence(news_text,filters='',lower=False)
            news_text_ori = " ".join(news_text_tok_ori)
            news_text_tok_num_ori = len(news_text_tok_ori)
#             news_text_tok_pre,rate_keras = text_preprcess(news_text,tokenizers='keras')
#             news_text_tok_pre,rate_nltk = text_preprcess(news_text,tokenizers='nltk')
            news_text_tok_pre,rate_spacy = text_preprcess(news_text,tokenizers='spacy')
            news_text_pre = " ".join(news_text_tok_pre)
        train_df.loc[count] = [id_num,cat,news_text_ori,news_text_tok_num_ori,news_text_pre,len(news_text_tok_pre)]
        count+=1
        break
    
train_df

以上deprecated

def text_preprcess2(text):
    '''
    Goal: filter out punctuations、words not in GloVe
    * cased
    * spacy
    * for RoBERTa + GloVe embedding
    Input: text string
    Return: preprcessed text string
    '''
    tokens = [i for i in text if i not in dil_filter] #濾掉符號
    tokens = sp_tokenizer(tokens)
    token_li = []
    for token in tokens:
        token_li.append(token.text)
    tokens_li = [i for i in token_li if i in glove_words] #只留有在glove的字
    tokens = list(filter(None, tokens_li))
    tokens_str = ' '.join(tokens)
    tokens_str = re.sub(dil," ",tokens_str) #在確認濾一次符號
    return tokens_str
    

train_df.loc[0,'text']

def basic_statistics(all_length):
    '''
    input: length list of elements e.g.[1,1,1,3,5,9,4,2,1,3,54,78,5...]
    output1: mean、std、mode、min、q1、median(q2)、q3、max、iqr、outlier、far out
    output2: statistics graph、10%~90% form
    '''
    stat_dict = {}
    stat_dict['mean'] = np.mean(all_length)
    stat_dict['std'] = np.std(all_length)
    stat_dict['mode'] = np.argmax(np.bincount(all_length))
    stat_dict['min'] = np.min(all_length)
    stat_dict['q1'] = np.quantile(all_length,0.25)
    stat_dict['median'] = np.quantile(all_length,0.5)
    stat_dict['q3'] = np.quantile(all_length,0.75)
    stat_dict['max'] = np.max(all_length)
    stat_dict['iqr'] = stat_dict['q3'] - stat_dict['q1']
    stat_dict['outlier'] = stat_dict['q3'] + 1.5*stat_dict['iqr']
    stat_dict['far_out'] = stat_dict['q3'] + 3*stat_dict['iqr']
    for i in [10,20,30,40,50,60,70,80,90,100]:
        stat_dict[str(i)+'%'] = np.percentile(all_length,i)
    return pd.DataFrame.from_dict(stat_dict,orient='index',columns=['length'])

tokens_num_li = train_df.preprocss_tokenize_num.tolist()
basic_statistics(tokens_num_li)

news_text_tok

news_list = next(os.walk(in_dir))[2]
news_list = [os.path.join(in_dir, f) for f in news_list]
news_list

# TEST

import torch

roberta = torch.hub.load('pytorch/fairseq', 'roberta.large')
roberta.eval() 

tokens_ori = "now let straight after nice long rant how people need personal responsibility economic social lives sudden 's radicals such me i guess responsible poor people 's lifestyles tell think poor people dumb think themselves there reasons disintegration family and support systems general nation 's poor somehow i think murphy janis the sane person 's list you want generation 's vaunted cultural revolution lasting change worse try socalled relevant values education hey like good idea time how know needed real education mean took granted the 's generation spoiled irresponsible the depression create mothers fathers determined kids want going overboard creating nation brats consider contrast famous events july apollo woodstock which group large numbers people feed reverted cultural level primitives defecation public etc and group assembled took care itself dispersed damage deaths large numbers drug problems was n't woodstock called biggest parking lot history they rejected society went nature parent 's cars"

# tokens_ori = train_df.loc[3,'preprocess_text']#'how\'s are you'#
tokens = roberta.encode(tokens_ori)
# assert tokens.tolist() == [0, 31414, 232, 328, 2]
# assert roberta.decode(tokens) == 'Hello world!'
print(len(tokens) , tokens)
roberta.decode(tokens)
doc = roberta.extract_features_aligned_to_words(tokens_ori)
print(len(doc))
for tok in doc:
    print('{:10}{} (...)'.format(str(tok), tok.vector[:5]))

tokens = 'I said, "hello RoBERTa."'
tokens = sp_tokenizer(tokens)
token_li = []
for token in tokens:
    token_li.append(token.text)
len(token_li) , token_li

last_layer_features = roberta.extract_features(tokens)
print(last_layer_features.size())# == torch.Size([1, 5, 1024])

all_layers = roberta.extract_features(tokens, return_all_hiddens=True)
assert len(all_layers) == 25
assert torch.all(all_layers[-1] == last_layer_features)

last_layer_features.detach().numpy().shape

from pytorch_transformers import RobertaTokenizer

tokenizer = RobertaTokenizer.from_pretrained('roberta-large-mnli')

len(tokenizer.encode(train_df.loc[0,'text']))
