==> 2.PrepareData.ipynb <==
import os, sys, glob, csv
# os.environ["CUDA_DEVICE_ORDER"]="PCI_BUS_ID"
# The GPU id to use, usually either "0" or "1"
os.environ["CUDA_VISIBLE_DEVICES"]="0" 
import numpy as np
import pandas as pd
import shutil
import pickle
import tqdm
from tqdm import tqdm
import xlsxwriter
import random
from random import shuffle
from math import log, floor
import re
import collections
from collections import Counter
import string
import unicodedata as udata
# import pause, datetime
import matplotlib.pyplot as plt
import seaborn as sns
from pathlib import Path
from distutils.dir_util import copy_tree
import sklearn
from sklearn.metrics import *
import itertools as it
import nltk
from nltk.corpus import stopwords
from nltk.tokenize import wordpunct_tokenize
from nltk import word_tokenize
from nltk.stem.porter import *
from nltk.tokenize import RegexpTokenizer
from collections import Counter
import functools
import spacy
import swifter
from pandarallel import pandarallel

import tensorflow.keras.preprocessing.text as T
from tensorflow.keras.utils import to_categorical
import torch
import dask.dataframe as dd
from dask.multiprocessing import get

import gc

read data

train_df = pd.read_csv('./data/20news-bydate-v3/train_df_1.csv',index_col=False)
train_df = train_df.drop(['Unnamed: 0'],axis=1)
valid_df = pd.read_csv('./data/20news-bydate-v3/valid_df_1.csv',index_col=False)
valid_df = valid_df.drop(['Unnamed: 0'],axis=1)
test_df = pd.read_csv('./data/20news-bydate-v3/test_df_1.csv',index_col=False)
test_df = test_df.drop(['Unnamed: 0'],axis=1)
train_df

category encode

categories = set(train_df.category.tolist())
print(len(categories))
cat_encode = {}
for i,cat in enumerate(categories):
    cat_encode[cat] = i
pickle.dump(obj=cat_encode,file=open('./data/20news-bydate-v3/cat_encode_dict.pkl','wb'))
cat_encode

def convert_cat(serie):
    '''
    trainY轉換為ID: 0~19 (20類)
    '''
    return cat_encode[serie]

train_df['cat_id'] = train_df.category.apply(convert_cat)
valid_df['cat_id'] = valid_df.category.apply(convert_cat)
test_df['cat_id'] = test_df.category.apply(convert_cat)
train_df

roberta = torch.hub.load('pytorch/fairseq', 'roberta.large')
roberta.cuda() #gpu
# roberta.cpu()
roberta.eval() 


# def tokenize_text(preprocess_text,max_length=481):
#     '''
#     利用RoBERTa tokenizer轉換 (太耗記憶體)
#     '''
#     roberta_pad = [0]*1024
#     doc = roberta.extract_features_aligned_to_words(preprocess_text)
#     roberta_emb = []
#     tokens_li = []
#     for tok in doc:
#         roberta_emb.append(tok.vector)
#         tokens_li.append(str(tok))
#     if len(roberta_emb)<max_length:
#         for _ in range(max_length-len(roberta_emb)):
#             roberta_emb.append(roberta_pad)
#     return np.array(roberta_emb) , tokens_li

# preprcoess_text_li = train_df.preprocess_text.tolist()
# roberta_pad = [0]*1024
# max_length = 498 #改
# train_df['roberta_emb'] = 0
# train_df['roberta_tok'] = 0
# # train_df['glove_emb'] = 0

# train_df=train_df.astype(object)

# glove = pd.read_table('./data/preprocess/glove.840B.300d.txt', sep=" ", index_col=0, header=None, quoting=csv.QUOTE_NONE)
# def glove_convert(tok_li):
#     '''
#     Deprecated
#     Input: list of tokens by RoBERTa
#     Return: np array of max_length*300dim
#     '''
#     max_length = 498 #改
#     glove_pad = [0]*300
#     glove_emb = []
#     for token in tok_li:
#         glove_emb.append(glove.loc[token].values)
#     if len(glove_emb)<max_length:
#         for _ in range(max_length-len(glove_emb)):
#             glove_emb.append(glove_pad)
#     try:
#         assert len(glove_emb) == max_length
#     except:
#         print("maxlength ERR(now,ori):",len(glove_emb) , max_length)
#     return np.array(glove_emb)



def transform1(train_df):
    preprcoess_text_li = train_df.preprocess_text.tolist()
    roberta_pad = [0]*1024
    max_length = 512 #改
    train_df['roberta_emb'] = 0
    train_df['roberta_tok'] = 0
    # train_df['glove_emb'] = 0

    train_df=train_df.astype(object)

    roberta_norm = [] # for dim norm
    norm_len = [] # for dim norm

    for i,text in tqdm(enumerate(preprcoess_text_li)):
        doc = roberta.extract_features_aligned_to_words(text)
        roberta_emb = []
        tokens_li = []

        for tok in doc:
            roberta_emb.append(tok.vector.cpu().data.numpy())
            tokens_li.append(str(tok))
        try:        
            assert len(tokens_li) == len(roberta_emb)
        except:
            print('AsserionERR: ',len(tokens_li), len(roberta_emb))
        roberta_norm.extend(roberta_emb) # for dim norm
        norm_len.append(len(roberta_emb))
        if len(roberta_emb)<max_length:
            for _ in range(max_length-len(roberta_emb)):
                roberta_emb.append(roberta_pad)
        train_df.loc[i,'roberta_emb'] = np.array(roberta_emb)
        train_df.loc[i,'roberta_tok'] = tokens_li
    #     train_df.loc[i,'glove_emb'] = glove_convert(tokens_li) #法1
        try:
            assert len(np.array(roberta_emb)) == (max_length)
        except:
            print('max length err(now max length,set max length):',len(np.array(roberta_emb)),(max_length))
    return train_df,roberta_norm,norm_len

train_df,train_roberta_norm,train_norm_len = transform1(train_df)

valid_df,valid_roberta_norm,valid_norm_len = transform1(valid_df)
test_df,test_roberta_norm,test_norm_len = transform1(test_df)
gc.collect()

embedding layer init matrix
* initial weight
* tokens #

# embeddings_index = {}
# f = open('./data/preprocess/glove.840B.300d.txt','r')
# for line in f:
#     values = line.split()
#     word = values[0]
#     coefs = np.asarray(values[1:], dtype='float32')
#     embeddings_index[word] = coefs
# f.close()

# print('Found %s word vectors.' % len(embeddings_index))


glove = pd.read_table('./data/preprocess/glove.840B.300d.txt', sep=" ", index_col=0, header=None, quoting=csv.QUOTE_NONE)
glove.index = glove.index.astype('str')
glove_words = glove.index.tolist()
print('all_iters#:',len(glove_words))
glove_index = {}
embedding_matrix = np.zeros((len(glove_words) + 1, 300))


for i,word in tqdm(enumerate(glove_words)):
    glove_index[word] = i+1
    try:
        embedding_matrix[i+1] = glove.loc[str(word)].values
    except KeyError:
        continue
    except ValueError:
        embedding_matrix[i+1] = glove.loc[str(word)].values[0]


pickle.dump(obj=glove_index,file=open('./data/20news-bydate-v3/glove_index.pkl','wb'))
gc.collect()
pickle.dump(obj=embedding_matrix,file=open('./data/20news-bydate-v3/embedding_matrix.pkl','wb'),protocol=4)

def tok2int(tok_li):
    '''
    Input: tok_list
    Return: int_list
    '''
    max_length = 512 #改
    int_pad = 0
    int_list = []
    for tok in tok_li:
        int_list.append(glove_index[str(tok)])
    assert len(tok_li) == len(int_list)
    if len(int_list)<max_length:
        for _ in range(max_length-len(tok_li)):
            int_list.append(int_pad)
    try:
        assert len(int_list) == max_length
    except:
        print("maxlength ERR(now,ori):",len(int_list) , max_length)            
    return int_list
def transform2(train_df):
    train_df['glove_int'] = 0
    train_df=train_df.astype(object)
    train_df['glove_int'] = train_df['roberta_tok'].swifter.apply(tok2int)
    return train_df

# train_df.to_csv('./data/20news-bydate-v3/train_df.csv',index=False)
# train_df

train_df = transform2(train_df)
pickle.dump(obj=train_df,file=open('./data/20news-bydate-v3/train_df_2.pkl','wb'))
# train_df.to_csv('./data/20news-bydate-v3/train_df.csv',index=False)
valid_df = transform2(valid_df)
pickle.dump(obj=valid_df,file=open('./data/20news-bydate-v3/valid_df_2.pkl','wb'))
# valid_df.to_csv('./data/20news-bydate-v3/valid_df.csv',index=False)
test_df = transform2(test_df)
pickle.dump(obj=test_df,file=open('./data/20news-bydate-v3/test_df_2.pkl','wb'))
# test_df.to_csv('./data/20news-bydate-v3/test_df.csv',index=False)
try:
    del embedding_matrix
    del glove_index
except:
    pass
gc.collect()

### dimension normalize
* roberta_emb

roberta_norm_train = np.array(train_roberta_norm)
print(roberta_norm_train.shape, len(train_norm_len))
mean = np.mean(roberta_norm_train,axis=0)
std = np.std(roberta_norm_train,axis=0)
pickle.dump(obj=mean,file=open('./data/20news-bydate-v3/dimNorm_mean.pkl','wb'))
pickle.dump(obj=std,file=open('./data/20news-bydate-v3/dimNorm_std.pkl','wb'))
mean.shape , std.shape

def transform3(train_df):
    all_train_roberta_emb = train_df.roberta_emb.tolist()
    all_train_roberta_emb = np.array(all_train_roberta_emb)
    print(all_train_roberta_emb.shape)
    return all_train_roberta_emb
# all_train_roberta_emb = all_train_roberta_emb.reshape(-1,1024)
# all_train_roberta_emb.shape

gc.collect()
all_train_roberta_emb = transform3(train_df)

# # del all_train_roberta_emb 
# del glove

# del valid_emb_norm
# del test_emb_norm
# import gc
# gc.collect()

def normalize_dim(emb,length,mean=mean,std=std):
    sent_emb = (emb[:length,:] - mean)/std
    final_emb = np.concatenate([sent_emb, emb[length:,:]],axis=0)
    return final_emb

# del roberta_norm_train
# del train_roberta_norm
glove = None
del glove
del all_valid_roberta_emb
del all_test_roberta_emb
gc.collect()

try:
    glove_words = None
    glove = None
    roberta_norm_train = None
    train_roberta_norm = None
    del roberta_norm_train
    del train_roberta_norm
    del glove
    del glove_words
except:
    pass
gc.collect()
train_emb_norm = [] 
for emb,length in tqdm(zip(all_train_roberta_emb,train_norm_len)):
    emb_norm = normalize_dim(emb,length)
    train_emb_norm.append(emb_norm)
all_train_roberta_emb = None
del all_train_roberta_emb
gc.collect()
train_emb_norm = np.array(train_emb_norm)
print(train_emb_norm.shape)
gc.collect()
pickle.dump(obj=train_emb_norm,file=open('./data/20news-bydate-v3/trainXr.pkl','wb'),protocol=4)

all_valid_roberta_emb = transform3(valid_df)
valid_emb_norm = []
for emb,length in tqdm(zip(all_valid_roberta_emb,valid_norm_len)):
    emb_norm = normalize_dim(emb,length)
    valid_emb_norm.append(emb_norm)
train_emb_norm = None
del all_valid_roberta_emb
del train_emb_norm
gc.collect()
valid_emb_norm = np.array(valid_emb_norm)
print(valid_emb_norm.shape)
pickle.dump(obj=valid_emb_norm,file=open('./data/20news-bydate-v3/validXr.pkl','wb'),protocol=4)

all_test_roberta_emb = transform3(test_df)
test_emb_norm = []
for emb,length in tqdm(zip(all_test_roberta_emb,test_norm_len)):
    emb_norm = normalize_dim(emb,length)
    test_emb_norm.append(emb_norm)
all_test_roberta_emb = None
del all_test_roberta_emb
gc.collect()
test_emb_norm = np.array(test_emb_norm)
print(test_emb_norm.shape)
pickle.dump(obj=test_emb_norm,file=open('./data/20news-bydate-v3/testXr.pkl','wb'),protocol=4)

valid_emb_norm = None
test_emb_norm = None
del test_emb_norm
del valid_emb_norm
gc.collect()
train_int = train_df.glove_int.tolist()
train_int = np.array(train_int)
print(train_int.shape)
pickle.dump(obj=train_int,file=open('./data/20news-bydate-v3/trainXg.pkl','wb'))

valid_int = valid_df.glove_int.tolist()
valid_int = np.array(valid_int)
print(valid_int.shape)
pickle.dump(obj=valid_int,file=open('./data/20news-bydate-v3/validXg.pkl','wb'))


test_int = test_df.glove_int.tolist()
test_int = np.array(test_int)
print(test_int.shape)
pickle.dump(obj=test_int,file=open('./data/20news-bydate-v3/testXg.pkl','wb'))

gc.collect()
train_ans = train_df.cat_id.tolist()
train_ans = to_categorical(train_ans)
pickle.dump(obj=train_ans,file=open('./data/20news-bydate-v3/trainY.pkl','wb'))
print(train_ans.shape)

valid_ans = valid_df.cat_id.tolist()
valid_ans = to_categorical(valid_ans)
pickle.dump(obj=valid_ans,file=open('./data/20news-bydate-v3/validY.pkl','wb'))
print(valid_ans.shape)

test_ans = test_df.cat_id.tolist()
test_ans = to_categorical(test_ans)
pickle.dump(obj=test_ans,file=open('./data/20news-bydate-v3/testY.pkl','wb'))
print(test_ans.shape)

filter pseudo ans
* norm_len個1e-8配上0，補到512

def create_dummy(len_li,max_len=512 ): #改
    '''
    Input: lens of list
    Return: dummy np
    '''
    result_li = []
    for length in len_li:
        front = np.full((length,),1e-8)
        back = np.full((max_len-length),0)
        result = np.concatenate((front,back),axis=0)
        result_li.append(result)
    return np.array(result_li)

train_pseudo_ans = create_dummy(train_norm_len)
valid_pseudo_ans = create_dummy(valid_norm_len)
test_pseudo_ans = create_dummy(test_norm_len)
pickle.dump(obj=(train_pseudo_ans,valid_pseudo_ans,test_pseudo_ans),
            file=open('./data/20news-bydate-v3/Yf.pkl','wb'))

keras to_categorical decode:

def decode(datum):
    return np.argmax(datum)
for i in range(encoded_data.shape[0]):
    datum = encoded_data[i]
    print('index: %d' % i)
    print('encoded datum: %s' % datum)
    decoded_datum = decode(encoded_data[i]) # decode method
    print('decoded datum: %s' % decoded_datum) # decode number
    print()

# Deprecated & Test
* Playground

法2
* deprecated

glove = pd.read_table('./data/preprocess/glove.840B.300d.txt', sep=" ", index_col=0, header=None, quoting=csv.QUOTE_NONE)
# train_df['glove_emb'] = 0
# train_df=train_df.astype(object)

def vec(w):
    return glove.loc[w].values

def glove_convert(tok_li):
    '''
    Input: list of tokens by RoBERTa
    Return: np array of max_length*300dim
    '''
    max_length = 512 #改
    glove_pad = [0]*300
    glove_emb = []
    for token in tok_li:
        glove_emb.append(glove.loc[token].values)
    if len(glove_emb)<max_length:
        for _ in range(max_length-len(glove_emb)):
            glove_emb.append(glove_pad)
    try:
        assert len(glove_emb) == max_length
    except:
        print("maxlength ERR(now,ori):",len(glove_emb) , max_length)
    return np.array(glove_emb)


train_df['glove_emb'] = train_df['roberta_tok'].swifter.apply(glove_convert)
train_df

# ddata = dd.from_pandas(train_df, npartitions=70)
# res = ddata.map_partitions(lambda df: df['roberta_tok'].apply(glove_convert).compute(get=get))
# res

# train_df['glove_emb'] = 0
# train_df=train_df.astype(object)
# pandarallel.initialize(shm_size_mb=8192) #shm_size_mb=8192,progress_bar=True
# train_df['glove_emb'] = train_df['roberta_tok'].parallel_apply(glove_convert) #swifter.apply
# train_df

train_df

train_df=train_df.astype(object)
# train_df['roberta_emb'] = train_df.roberta_emb.astype(object)
train_df.loc[i,'roberta_emb'] = np.array(roberta_emb)

train_df.roberta_emb[0].shape

type(train_df.id[0])

train_df['roberta_emb'], train_df['roberta_tok'] = zip(*train_df['preprocess_text'].map(tokenize_text))
train_df

train_df

train_df = pd.read_csv('./data/20news-bydate-v3/train_df.csv')
valid_df = pd.read_csv('./data/20news-bydate-v3/valid_df.csv')
test_df = pd.read_csv('./data/20news-bydate-v3/test_df.csv')
train_df

train_len_li = train_df.roberta_tok.tolist()
train_len_li = [len(x.split("', '")) for x in train_len_li]

train_len_li[0]

train_df.glove_int[0].split(', ')[164]

train_df.roberta_emb[0]
